publisher/requirements.txt
pika==1.2.1
ipdb
python-dotenv